+++ lib/models/charity.dart
@@ -3,13 +3,15 @@ class Charity {
   String id;
   String description;
   String website;
+  String imagePath;
   String paymentInfo;
 
-  Charity(String name, String id, String description, String website, String paymentInfo) {
+  Charity(String name, String id, String description, String website, String imagePath, String paymentInfo) {
     this.name = name;
     this.id = id;
     this.description = description;
     this.website = website;
+    this.imagePath = imagePath;
     this.paymentInfo = paymentInfo;
   }
 }